Created Categories App. Moved models to to Category Models. Moved Urls to Category Urls. Moved Views to Category Views
