+++ scripted_be/settings.py
@@ -36,6 +36,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'users',
+    'category',
 ]
 
 MIDDLEWARE = [
