+++ scripted_be/urls.py
@@ -8,8 +8,11 @@ from rest_framework_simplejwt.views import TokenRefreshView
 from users.urls import users_router
 from users.views import CustomTokenObtainPairView
 
+from category.urls import category_router
+
 router = routers.SimpleRouter(trailing_slash=False)
 router.registry.extend(users_router.registry)
+router.registry.extend(category_router.registry)
 
 api_urlpatterns = [
     path('', include(router.urls)),
