+++ users/urls.py
@@ -1,7 +1,6 @@
 from rest_framework.routers import DefaultRouter
 
-from users.views import SignUpViewSet, CategoriesViewSet
+from users.views import SignUpViewSet
 
 users_router = DefaultRouter()
 users_router.register(r'signup', SignUpViewSet, basename='signup')
-users_router.register(r'category', CategoriesViewSet, basename='category')
