+++ users/views.py
@@ -3,7 +3,6 @@ from rest_framework.response import Response
 from rest_framework import viewsets
 from rest_framework_simplejwt.views import TokenObtainPairView
 
-from users.models import Category
 from users.serializers import UserSerializer, CustomTokenObtainPairSerializer, CategorySerializer
 
 
@@ -27,17 +26,3 @@ class CustomTokenObtainPairView(TokenObtainPairView):
     serializer_class = CustomTokenObtainPairSerializer
 
 
-class CategoriesViewSet(viewsets.ViewSet):
-    """
-    ViewSet for categories.
-    """
-
-    def list(self, request):
-        categories = Category.objects.all()
-        serializer = CategorySerializer(categories, many=True)
-        return Response(serializer.data)
-
-    def retrieve(self, request, pk=None):
-        category = Category.objects.get(id=pk)
-        serializer = CategorySerializer(category)
-        return Response(serializer.data)
